showBookFinal
