+++ public/assets/css/mobile.css
@@ -5,6 +5,7 @@
 
 body {
     background-color: var(--color-background);
+    margin-bottom: 25%;
 }
 
 h1 {
@@ -81,18 +82,18 @@ div.icones {
 
 #details {
     display: none;
-    text-align: center;
-    line-height: 4rem;
-    margin: 2rem;
-    font-size: 1.7rem;
+    text-align:unset; 
+    line-height: 2.5rem;
+    margin: 1.5rem;
+    font-size: 1.5rem;
     color: var(--color-main);
 }
 
 #review {
     display: none;
-    text-align: center;
-    line-height: 3rem;
-    margin: 2rem;
-    font-size: 1.7rem;
+    text-align: start;
+    line-height: 2rem;
+    margin: 1.5rem;
+    font-size: 1.5rem;
     color: var(--color-main);
 }
