+++ public/index.php
@@ -1,5 +1,6 @@
 <?php
 
+
 require_once __DIR__ . '/../vendor/autoload.php';
 
 require_once __DIR__ . '/../config/debug.php';
