+++ src/Model/BookEditorManager.php
@@ -44,4 +44,18 @@ class BookEditorManager extends AbstractManager
 
         return $statement->fetch(PDO::FETCH_ASSOC);
     }
+
+    public function selectOneById(int $id): array|false
+    {
+        $statement = $this->pdo->prepare('SELECT b.title, b.written_at, 
+        be.cover, be.synopsis, be.isbn, be.nb_pages, 
+        r.created_at, r.note, r.reading_time, r.difficulty, r.opinion, 
+        a.firstname, a.lastname, e.label, g.label 
+        FROM ' . static::TABLE . ' AS be JOIN book b JOIN review r 
+        JOIN author a JOIN editor e JOIN genre g ON b.id=be.book_id WHERE b.id=:id');
+        $statement->bindValue('id', $id, PDO::PARAM_INT);
+        $statement->execute();
+
+        return $statement->fetch();
+    }
 }
