+++ src/View/Book/_navbar.html.twig
@@ -8,7 +8,7 @@
             </a>
         </div>
         <div class="btn add">
-            <a href="">
+            <a href="formAdd.html.twig">
                 <img class="icon-mobile" src="/assets/images/navbar/icon-add.png" alt="Add button">
                 <img class="icon-desktop" src="/assets/images/navbar/icon-add-colored.png" alt="Add button">
             </a>
