+++ src/View/Book/show.html.twig
@@ -25,7 +25,7 @@
   </div>
   <div id="note">
     <p> 
-      note : {{ book.note }}/10
+      note globale : {{ book.note }}/5
     </p>
   </div>
 </div>
@@ -52,13 +52,12 @@
   </p>
 </DIV>
 <div id="review">
-  <p>Date d'ajout : <strong>{{book.created_at }}</strong></p>
   <p>Difficulté : <strong>{{ book.difficulty }}/5</strong></p>
   <p>Temps de lecture moyen : <strong>{{ book.reading_time }}h</strong></p>
+  <p>Date d'ajout : <strong>{{ book.created_at }}</strong></p>
   <p>Commentaires : <strong>{{ book.opinion }}</strong></p>
 </div>
 <div id="details">
-  <p> Langue : <strong>{{ book.langue }}</strong></p>
   <p>ISBN : <strong>{{ book.isbn }}</strong></p>
   <p>Genre : <strong>{{ book.label }}</strong></p>
   <p>Date de publication : <strong>{{ book.written_at }}</strong></p>
