+++ src/View/Home/index.html.twig
@@ -20,4 +20,4 @@
 	</div>
 	{% include 'Book/_navbar.html.twig' %}
 {% endblock %}
->>>>>>> dev
+
