+++ src/View/layout.html.twig
@@ -13,6 +13,7 @@
 
 		<link rel="stylesheet" href="https://maxst.icons8.com/vue-static/landings/line-awesome/line-awesome/1.3.0/css/line-awesome.min.css">
 		<link rel="stylesheet" href="/assets/css/style.css">
+		<link rel="stylesheet" href="/assets/css/mobile.css">
 		<link rel="stylesheet" href="/assets/css/desktop.css" media="(min-width: 1200px)"> {% block stylesheet %}{% endblock %}
 		</head>
 
@@ -31,5 +32,4 @@
 				<script src="/assets/js/navbar.js"></script>
 			{% endblock %}
 		</body>
-		>>>>>>> dev
 	</html>
